user domain expection filter implmented to handle user related error

+++ src/infrastructure/database/mongo/user/mongo.user.repository.ts
@@ -1,13 +1,12 @@
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { UserRepository } from 'src/modules/users/domain/user.repository';
 import { UserEntity } from 'src/modules/users/domain/user.entity';
 import { CreateUserDto } from 'src/modules/users/dto/create-user.dto';
 import { User, UserDocument } from 'src/infrastructure/database/mongo/schemas/user/user.schema';
 import { UserMapper } from 'src/modules/users/users.mapper';
-import { isUniqueConstraintViolation } from 'src/infrastructure/errors/db-error.mapper';
-import { EmailAlreadyExistsError } from 'src/modules/users/domain/errors/email-already-exits.error';
 import { UserNotFoundError } from 'src/modules/users/domain/errors/user-not-found.error';
+import { BadRequestException } from '@nestjs/common';
 
 export class MongoUserRepository implements UserRepository {
   constructor(
@@ -20,15 +19,15 @@ export class MongoUserRepository implements UserRepository {
       const user = await this.userModel.create(dto);
       return UserMapper.toEntity(user.toObject());
     } catch (err) {
-      if (isUniqueConstraintViolation(err)) {
-        throw new EmailAlreadyExistsError('email');
-      }
       throw err;
     }
   }
 
   async findById(id: string): Promise<UserEntity | null> {
     try {
+      if (!Types.ObjectId.isValid(id)) {
+        throw new BadRequestException('Invalid user ID format');
+      }
       const user = await this.userModel.findById(id);
       if (!user) {
         throw new UserNotFoundError(id);
@@ -38,14 +37,7 @@ export class MongoUserRepository implements UserRepository {
       throw err;
     }
   }
-  async findByEmail(email: string): Promise<UserEntity | null> {
-    try {
-      const user = await this.userModel.findOne({ email });
-      return user ? UserMapper.toPublic(user.toObject()) : null;
-    } catch (err) {
-      throw err;
-    }
-  }
+
   async findAll(): Promise<UserEntity[]> {
     try {
       const users = await this.userModel.find();
@@ -67,6 +59,9 @@ export class MongoUserRepository implements UserRepository {
   }
   async delete(id: string): Promise<void> {
     try {
+      if (!Types.ObjectId.isValid(id)) {
+        throw new BadRequestException('Invalid user ID format');
+      }
       await this.userModel.findByIdAndDelete(id);
     } catch (err) {
       throw err;
